+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [1.0.3] - 2019-10-19
+
+### Added
+
+- Support for `rustup set profile` command
+- Support for `--component` flag for the `rustup toolchain install` command
+
 ## [1.0.2] - 2019-10-14
 
 ### Changed